Shouldn't need a recompile to change window size.

+++ Source/Sample_D3D11LightProbe.cpp
@@ -103,7 +103,8 @@ void main(uint3 dispatchThreadId : SV_DispatchThreadID)
       Constants constants;
       constants.TransformClipToWorld =
           Invert(GetTransformSource()->GetTransformWorldToClip());
-      constants.WindowDimensions = {RENDERTARGET_WIDTH, RENDERTARGET_HEIGHT};
+      constants.WindowDimensions = {static_cast<float>(RENDERTARGET_WIDTH),
+                                    static_cast<float>(RENDERTARGET_HEIGHT)};
       device->GetID3D11DeviceContext()->UpdateSubresource(
           bufferConstants, 0, nullptr, &constants, 0, 0);
     }
